feat: change in the registration value of the habit

+++ app/src/main/java/space/algoritmos/habit_tracker/data/local/DatabaseHelper.kt
@@ -20,7 +20,7 @@ class DatabaseHelper(context: Context) : SQLiteOpenHelper(
                 name TEXT NOT NULL,
                 color TEXT,
                 tracking_mode TEXT,
-                goal REAL,
+                goal INT,
                 progress TEXT
             )
         """.trimIndent()
